+++ tests/testthat/testfunctions.R
@@ -26,10 +26,15 @@ has_internet <- function(){
     skip("No internet connection available. ")
   }
   if ("try-error" %in% c(
-    class(try(httr::headers(httr::HEAD(url = utils::URLencode("https://www.clinicaltrialsregister.eu/"))), silent = TRUE)),
-    class(try(httr::headers(httr::HEAD(url = utils::URLencode("https://clinicaltrials.gov"))),             silent = TRUE))
-
-    )
+    class(try(httr::headers(httr::HEAD(
+      url = utils::URLencode("https://clinicaltrials.gov"))),
+      silent = TRUE)),
+    class(try(httr::headers(httr::HEAD(
+      url = utils::URLencode("https://www.clinicaltrialsregister.eu/"),
+      config = httr::config(ssl_verifypeer = FALSE)
+      )),
+      silent = TRUE))
+  )
   ) {
     skip("One or more registers not available. ")
   }
@@ -40,17 +45,14 @@ has_mongo <- function(){
   # check server
   tmp <- getOption("warn")
   options("warn" = 2)
-  mongo_ok <- try({
-    capture.output(ctrMongo())
-  },
-  silent = TRUE)
+  # test
+  mongo_ok <- try(ctrdata:::ctrMongo(), silent = TRUE)
   # use test result
   options("warn" = tmp)
-  if (class(mongo_ok) == "try-error") {
-    skip("No suitable version password-free localhost mongodb accessible.")
+  if ("try-error" %in% class(mongo_ok)) {
+    skip("No password-free localhost mongodb accessible.")
   }
 }
-
 
 # helper function to check tool chain
 has_toolchain <- function(){
@@ -68,9 +70,6 @@ has_toolchain <- function(){
         ifelse(is.null(tmp), TRUE, tmp)
       }),
 
-      # this test for class is only needed to obtain a logical value
-      !(class(installMongoFindBinaries()) == "character"),
-
       na.rm = TRUE)
   }, silent = TRUE)
 
@@ -78,7 +77,6 @@ has_toolchain <- function(){
     skip("One or more tool chain applications are not available.")
   }
 }
-
 
 #### mongodb local password free access ####
 test_that("access to mongo db from R package", {
@@ -89,8 +87,7 @@ test_that("access to mongo db from R package", {
   coll <- "ThisNameSpaceShouldNotExistAnywhereInAMongoDB"
 
   # initialise = drop collections from mongodb
-  try(mongolite::mongo(collection = coll,                 db = "users")$drop(), silent = TRUE)
-  try(mongolite::mongo(collection = paste0(coll, "Keys"), db = "users")$drop(), silent = TRUE)
+  try(mongolite::mongo(collection = coll, url = "mongodb://localhost/users")$drop(), silent = TRUE)
 
   # test 1
   expect_message(dbQueryHistory(collection = coll),
@@ -99,61 +96,47 @@ test_that("access to mongo db from R package", {
 })
 
 
-#### mongodb access ####
-test_that("access to mongo db from command line", {
-
-  has_mongo()
-
-  # test 2
-  expect_message(installMongoFindBinaries(debug = TRUE),
-                 "mongo / mongoimport ")
-
-})
-
-
 #### remote mongodb ####
-test_that("retrieve data from registers", {
+test_that("remote mongo access", {
 
   has_internet()
   has_toolchain()
 
   # initialise
+  uri <- "mongodb+srv://7RBnH3BF@cluster0-b9wpw.mongodb.net/dbtemp"
   coll <- "ThisNameSpaceShouldNotExistAnywhereInAMongoDB"
 
   # brief testing of main functions
 
-  # test 3a
-  expect_message(suppressWarnings(
-    ctrLoadQueryIntoDb(
-      queryterm = "2012-004228-40",
-      register = "EUCTR",
-      url = "mongodb+srv://cluster0-b9wpw.mongodb.net/",
-      db = "dbtemp",
-      collection = "dbcoll",
-      username = "admin",
-      password = "admin")
-  ),
-  "First result page empty")
+  # note password is set as environment variable
+  # which is read by ctrdata main functions
 
-  # test 3b
-  dbQueryHistory(url = "mongodb+srv://cluster0-b9wpw.mongodb.net/",
-                 db = "dbtemp",
-                 collection = "dbcoll",
-                 username = "admin",
-                 password = "admin")
+  # test 2a
+  expect_equivalent(ctrLoadQueryIntoDb(
+    queryterm = "2010-024264-18",
+    register = "CTGOV",
+    uri = uri,
+    collection = coll)$n,
+    1L)
 
-  # test 3c
-  expect_warning(
-    dbFindFields(
-      namepart = "x6_date_on_which_this_record_was_first_entered_in_the_eudract_database",
-      url = "mongodb+srv://cluster0-b9wpw.mongodb.net/",
-      db = "dbtemp",
-      collection = "dbcoll",
-      username = "admin",
-      password = "admin"),
-    "")
+  # test 2b
+  expect_equivalent(ctrLoadQueryIntoDb(
+    queryterm = "2010-024264-18",
+    register = "EUCTR",
+    uri = uri,
+    collection = coll)$n,
+    6L)
 
-  # clean up is the end of script = drop collection from mongodb
+  # test 2c
+  expect_true(
+    length(
+      suppressWarnings(
+        dbFindFields(namepart = "date",
+                     uri = uri,
+                     collection = coll))) > 5)
+
+  # clean up
+  ctrdata:::ctrMongo(uri = uri, collection = coll)$drop()
 
 })
 
@@ -169,18 +152,18 @@ test_that("retrieve data from registers", {
   coll <- "ThisNameSpaceShouldNotExistAnywhereInAMongoDB"
 
   # test 3
-  expect_message(suppressWarnings(ctrLoadQueryIntoDb(
+  expect_equal(suppressWarnings(ctrLoadQueryIntoDb(
     queryterm = "query=NonExistingConditionGoesInHere",
     register = "EUCTR",
-    collection = coll)),
-    "First result page empty")
+    collection = coll)$n),
+    0L)
 
   # test 4
-  expect_error(suppressWarnings(ctrLoadQueryIntoDb(
+  expect_equal(suppressWarnings(ctrLoadQueryIntoDb(
     queryterm = "cond=NonExistingConditionGoesInHere",
     register = "CTGOV",
-    collection = coll)),
-    "No trials or number of trials could not be determined")
+    collection = coll)$n),
+    0L)
 
   # clean up is the end of script = drop collection from mongodb
 
@@ -204,14 +187,17 @@ test_that("retrieve data from register ctgov", {
     collection = coll),
     "Imported or updated 1 trial")
 
-
   ## create and test updatable query
 
   q <- paste0("https://clinicaltrials.gov/ct2/results?term=osteosarcoma&type=Intr&phase=0&age=0&lup_e=")
 
   # test 6
-  expect_message(ctrLoadQueryIntoDb(paste0(q, "12%2F31%2F2008"), collection = coll),
-                 "Imported or updated ")
+  expect_message(ctrLoadQueryIntoDb(
+    paste0(q, "12%2F31%2F2008"),
+    collection = coll,
+    debug = TRUE,
+    verbose = TRUE),
+    "Imported or updated ")
 
   # manipulate history to force testing updating
   # based on code in dbCTRUpdateQueryHistory
@@ -221,13 +207,15 @@ test_that("retrieve data from register ctgov", {
   # convert into json object
   json <- jsonlite::toJSON(list("queries" = hist))
   # update database
-  mongolite::mongo(collection = coll, db = "users")$update(query = '{"_id":{"$eq":"meta-info"}}',
-                                                           update = paste0('{ "$set" :', json, "}"),
-                                                           upsert = TRUE)
+  mongolite::mongo(collection = coll,
+                   url = "mongodb://localhost/users")$update(query = '{"_id":{"$eq":"meta-info"}}',
+                                                             update = paste0('{ "$set" :', json, "}"),
+                                                             upsert = TRUE)
 
   # test 7
-  expect_message(suppressWarnings(ctrLoadQueryIntoDb(querytoupdate = "last", collection = coll)),
-                 "Imported or updated")
+  expect_message(suppressWarnings(ctrLoadQueryIntoDb(
+    querytoupdate = "last", collection = coll)),
+    "Imported or updated")
 
   remove("hist", "json", "q")
 
@@ -244,21 +232,6 @@ test_that("retrieve data from register euctr", {
   # initialise
   coll <- "ThisNameSpaceShouldNotExistAnywhereInAMongoDB"
 
-  # test 8
-  expect_message(ctrLoadQueryIntoDb(
-    queryterm = "2010-024264-18",
-    register = "EUCTR",
-    collection = coll),
-    "Updated history")
-
-  # test 9
-  expect_message(suppressWarnings(
-    ctrLoadQueryIntoDb(
-      querytoupdate = "last",
-      collection = coll)),
-    "First result page empty")
-
-  ## forced slow import
   q <- paste0("https://www.clinicaltrialsregister.eu/ctr-search/search?query=",
               "neuroblastoma&status=completed&phase=phase-one&country=pl")
   # ctrGetQueryUrlFromBrowser(content = q)
@@ -266,9 +239,7 @@ test_that("retrieve data from register euctr", {
   # test 11
   expect_message(suppressWarnings(
     ctrLoadQueryIntoDb(q,
-                       collection = coll,
-                       debug = TRUE,
-                       verbose = FALSE)),
+                       collection = coll)),
     "Imported or updated")
 
   ## download without details
@@ -309,9 +280,9 @@ test_that("retrieve data from register euctr", {
   json <- jsonlite::toJSON(list("queries" = hist))
   # update database
   mongolite::mongo(collection = coll,
-                   db = "users")$update(query = '{"_id":{"$eq":"meta-info"}}',
-                                        update = paste0('{ "$set" :', json, "}"),
-                                        upsert = TRUE)
+                   url = "mongodb://localhost/users")$update(query = '{"_id":{"$eq":"meta-info"}}',
+                                                             update = paste0('{ "$set" :', json, "}"),
+                                                             upsert = TRUE)
 
   # test 14
   expect_message(
@@ -451,67 +422,80 @@ test_that("operations on database after download from register", {
   coll <- "ThisNameSpaceShouldNotExistAnywhereInAMongoDB"
 
   # test 28
-  expect_error(dbFindFields(namepart = c("onestring", "twostring"), collection = coll),
-               "Name part should have only one element.")
+  expect_error(dbFindFields(
+    namepart = c("onestring", "twostring"),
+    collection = coll),
+    "Name part should have only one element.")
 
   # test 29
-  expect_error(dbFindFields(namepart = list("onestring", "twostring"), collection = coll),
-               "Name part should be atomic.")
+  expect_error(dbFindFields(
+    namepart = list("onestring", "twostring"),
+    collection = coll),
+    "Name part should be atomic.")
 
   # test 30
-  expect_error(dbFindFields(namepart = "", collection = coll),
+  expect_error(dbFindFields(namepart = "",
+                            collection = coll),
                "Empty name part string.")
 
   # test 31
-  expect_message(dbFindFields(namepart = "date", collection = coll),
-                 "Returning first of ")
+  expect_type(dbFindFields(
+    namepart = "date",
+    collection = coll),
+    "character")
 
   # test 32
-  expect_equal(is.na(dbFindFields(namepart = "ThisNameShouldNotExistAnywhere", collection = coll)),
-               TRUE)
+  expect_equal(length(dbFindFields(
+    namepart = "ThisNameShouldNotExistAnywhere",
+    collection = coll)), 0L)
 
   # dbFindIdsUniqueTrials
 
   # test 33
-  expect_message(dbFindIdsUniqueTrials(collection = coll, preferregister = "EUCTR"),
-                 "Searching multiple country records")
+  expect_message(dbFindIdsUniqueTrials(
+    collection = coll,
+    preferregister = "EUCTR"),
+    "Searching multiple country records")
 
   # test 34
-  expect_message(dbFindIdsUniqueTrials(collection = coll, preferregister = "CTGOV"),
-                 "Returning keys")
+  expect_message(dbFindIdsUniqueTrials(
+    collection = coll,
+    preferregister = "CTGOV"),
+    "Returning keys")
 
   # test 35
-  expect_warning(dbFindIdsUniqueTrials(collection = coll, prefermemberstate = "3RD", include3rdcountrytrials = FALSE),
-                 "Preferred EUCTR version set to 3RD country trials, but include3rdcountrytrials was FALSE")
+  expect_warning(dbFindIdsUniqueTrials(
+    collection = coll,
+    prefermemberstate = "3RD",
+    include3rdcountrytrials = FALSE),
+    "Preferred EUCTR version set to 3RD country trials, but include3rdcountrytrials was FALSE")
 
 
   # dbGetFieldsIntoDf
 
-  # test 36a (temporary)
-  expect_warning(dbGetVariablesIntoDb(fields = "x6_date_on_which_this_record_was_first_entered_in_the_eudract_database",
-                                      collection = coll)[1, ],
-                 "'dbGetVariablesIntoDb' is deprecated.")
-
   # test 36
-  expect_error(dbGetFieldsIntoDf(fields = "ThisDoesNotExist", collection = coll),
-               "For field: ThisDoesNotExist no data could be extracted")
+  expect_error(dbGetFieldsIntoDf(
+    fields = "ThisDoesNotExist",
+    collection = coll),
+    "For field: ThisDoesNotExist no data could be extracted")
 
   # test 37
-  expect_error(dbGetFieldsIntoDf(fields = "", collection = coll),
-               "'fields' contains empty elements")
+  expect_error(dbGetFieldsIntoDf(
+    fields = "",
+    collection = coll),
+    "'fields' contains empty elements")
 
   # test 38
-  expect_error(dbGetFieldsIntoDf(fields = list("ThisDoesNotExist"), collection = coll),
-               "Input should be a vector of strings of field names.")
+  expect_error(dbGetFieldsIntoDf(
+    fields = list("ThisDoesNotExist"),
+    collection = coll),
+    "Input should be a vector of strings of field names.")
 
 
   # clean up = drop collections from mongodb
 
   # test 38
-  expect_equivalent (mongolite::mongo(collection = coll,                 db = "users")$drop(), TRUE)
-
-  # test 40
-  expect_equivalent (mongolite::mongo(collection = paste0(coll, "Keys"), db = "users")$drop(), TRUE)
+  expect_equivalent(mongolite::mongo(collection = coll, url = "mongodb://localhost/users")$drop(), TRUE)
 
 })
 
@@ -536,18 +520,22 @@ test_that("operations on database for deduplication", {
   # ctrLoadQueryIntoDb(queryterm = "2014-005674-11", register = "EUCTR", collection = coll) # this is 3rd country only
   # ctrLoadQueryIntoDb(queryterm = "2016-002347-41", register = "EUCTR", collection = coll) # in eu and 3rd country
 
-  ctrLoadQueryIntoDb(queryterm = "NCT00134030 OR NCT01516580 OR NCT00025597",
-                     register = "CTGOV", collection = coll)
+  ctrLoadQueryIntoDb(
+    queryterm = "NCT00134030 OR NCT01516580 OR NCT00025597",
+    register = "CTGOV",
+    collection = coll)
 
-  ctrLoadQueryIntoDb(queryterm = "2010-019224-31 OR 2004-000242-20 OR 2005-000915-80 OR 2014-005674-11 OR 2016-002347-41",
-                     register = "EUCTR", collection = coll)
+  ctrLoadQueryIntoDb(
+    queryterm = "2010-019224-31 OR 2004-000242-20 OR 2005-000915-80 OR 2014-005674-11 OR 2016-002347-41",
+    register = "EUCTR",
+    collection = coll)
 
-    # test combinations of parameters
+  # test combinations of parameters
 
   # test 41
   tmp <- dbFindIdsUniqueTrials(collection = coll)
-  expect_true(all.equal(tmp, c("2004-000242-20-GB", "2005-000915-80-GB", "2010-019224-31-GB", "2014-005674-11-3RD",
-                               "2016-002347-41-GB", "NCT00025597"),
+  expect_true(all.equal(tmp, c("2004-000242-20-GB", "2005-000915-80-GB", "2010-019224-31-GB",
+                               "2014-005674-11-3RD", "2016-002347-41-GB", "NCT00025597"),
                         check.attributes = FALSE))
 
   # test 42
@@ -584,10 +572,7 @@ test_that("operations on database for deduplication", {
   # clean up = drop collections from mongodb
 
   # test 47
-  expect_equivalent (mongolite::mongo(collection = coll,                 db = "users")$drop(), TRUE)
-
-  # test 48
-  expect_equivalent (mongolite::mongo(collection = paste0(coll, "Keys"), db = "users")$drop(), TRUE)
+  expect_equivalent(mongolite::mongo(collection = coll, url = "mongodb://localhost/users")$drop(), TRUE)
 
 })
 
@@ -657,7 +642,9 @@ test_that("annotate queries", {
 #### df operations ####
 test_that("operations on data frame", {
 
-  df <- data.frame("var1" = 1:3, "var2" = 2:4, stringsAsFactors = FALSE)
+  df <- data.frame("var1" = 1:3,
+                   "var2" = 2:4,
+                   stringsAsFactors = FALSE)
 
   statusvalues <- list("Firstvalues" = c("12", "23"),
                        "Lastvalue"   = c("34"))
@@ -669,21 +656,25 @@ test_that("operations on data frame", {
                "Need a data frame as input.")
 
   # test 54
-  expect_message(dfMergeTwoVariablesRelevel(df = df, colnames = c("var1", "var2")),
+  expect_message(dfMergeTwoVariablesRelevel(df = df,
+                                            colnames = c("var1", "var2")),
                  "Unique values returned: 12, 23, 34")
 
   # test 55
-  expect_is     (dfMergeTwoVariablesRelevel(df = df, colnames = c("var1", "var2")),
-                 "character")
+  expect_is(dfMergeTwoVariablesRelevel(df = df,
+                                       colnames = c("var1", "var2")),
+            "character")
 
   # test 56
-  expect_message(dfMergeTwoVariablesRelevel(df = df, colnames = c("var1", "var2"),
+  expect_message(dfMergeTwoVariablesRelevel(df = df,
+                                            colnames = c("var1", "var2"),
                                             levelslist = statusvalues),
                  "Unique values returned: Firstvalues, Lastvalue")
 
   # test 57
-  expect_error(dfMergeTwoVariablesRelevel(df = df, colnames = 1:3),
-                 "Please provide exactly two column names.")
+  expect_error(dfMergeTwoVariablesRelevel(df = df,
+                                          colnames = 1:3),
+               "Please provide exactly two column names.")
 
 
 })